feat: restructure SplashWaveEffect component and related shaders for improved organization and maintainability
